requirements_screen.txt
pywin32==306
psutil==5.9.8
mss==9.0.1
Pillow==10.3.0
requests==2.31.0
